Fix rap/elf archive file load
When loading an object file in an archive file for the first time, RTL
complains that loader can not be found. It is because offset is bigger
that cache->file_size the first time when file_size is still -1, while
offset is positive because reading the archive file. This patch fixes
ths problem. Flush the file_size to 0 but not -1, because file_size is
unsigned type.